Compositor PageFault Fixed, Multiboot Memory map added!

+++ src/Kernel/Atomix.Kernel_H/gui/Compositor.cs
@@ -147,7 +147,7 @@ namespace Atomix.Kernel_H.gui
                             var xNewWindow = new Window(ClientID) { Width = WindowWidth, Height = WindowHeight };
 
                             string HashString = xNewWindow.HashString;
-                            xNewWindow.Buffer = SHM.Obtain(HashString, (int)(WindowHeight * WindowWidth * VBE.BytesPerPixel), true);
+                            xNewWindow.Buffer = SHM.Obtain(HashString, (WindowHeight * WindowWidth * VBE.BytesPerPixel), true);
                             
                             //Yes! we overwrite this buffer because we have no further refernce to this
                             Helper.CreateNewWindowMessage(compositor_packet, WindowWidth, WindowHeight, HashString);
